adds new zome called users
